Storage optimizations and refactoring
